Create shared API key status
