+++ packages/maas/frontend/src/app/pages/keys-and-subs/apiKeys/allKeys/ApiKeysTableRow.tsx
@@ -1,35 +1,10 @@
 import * as React from 'react';
 import { ActionsColumn, Td, Tr } from '@patternfly/react-table';
-import { capitalize, Label, Popover } from '@patternfly/react-core';
-import type { LabelProps } from '@patternfly/react-core';
-import {
-  BanIcon,
-  CheckCircleIcon,
-  MinusCircleIcon,
-  OutlinedClockIcon,
-  OutlinedQuestionCircleIcon,
-} from '@patternfly/react-icons';
 import TableRowTitleDescription from '@odh-dashboard/internal/components/table/TableRowTitleDescription';
 import { APIKey, APIKeyDisplayStatus, SubscriptionDetail } from '~/app/types/api-key';
+import ApiKeyStatusLabel from '~/app/pages/keys-and-subs/apiKeys/ApiKeyStatusLabel';
 import { ApiKeyColumn } from './columns';
 import SubscriptionCell from './SubscriptionCell';
-
-const getApiKeyStatusProps = (
-  status: APIKeyDisplayStatus,
-): { icon: React.ReactNode; status?: LabelProps['status']; variant?: LabelProps['variant'] } => {
-  switch (status) {
-    case 'active':
-      return { icon: <CheckCircleIcon />, status: 'success' };
-    case 'inactive':
-      return { icon: <MinusCircleIcon />, variant: 'filled' };
-    case 'expired':
-      return { icon: <OutlinedClockIcon /> };
-    case 'revoked':
-      return { icon: <BanIcon />, status: 'danger' };
-    default:
-      return { icon: <OutlinedQuestionCircleIcon /> };
-  }
-};
 
 const formatDate = (dateString?: string, fallback = '—'): string => {
   if (!dateString) {
@@ -68,25 +43,7 @@ const cellRenderers: Record<string, CellRenderer> = {
 
   status: ({ apiKey, isInactive }) => {
     const displayStatus = getDisplayStatus(apiKey, isInactive);
-    const { variant, ...labelProps } = getApiKeyStatusProps(displayStatus);
-    const label = (
-      <Label variant={variant ?? 'outline'} {...labelProps}>
-        {capitalize(displayStatus)}
-      </Label>
-    );
-
-    if (displayStatus !== 'inactive') {
-      return label;
-    }
-
-    return (
-      <Popover
-        headerContent="Subscription unavailable"
-        bodyContent="The subscription this key was created for has been deleted or is not ready. The key itself has not been revoked, but it cannot authenticate requests until the subscription is restored."
-      >
-        <span style={{ cursor: 'pointer' }}>{label}</span>
-      </Popover>
-    );
+    return <ApiKeyStatusLabel status={displayStatus} showInactivePopover />;
   },
 
   subscription: ({ apiKey, subscriptionDetail }) => (
